pyproject.toml
[tool.poetry]
name = "test-tasks"
version = "0.1.0"
description = ""
authors = ["DL <dmitriy.luzanovsky@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.5"
djangorestframework = "^3.15.2"
django-q2 = "^1.7.6"
requests = "^2.32.3"
drf-yasg = "^1.21.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
